Rewrite the darkening logic using fontify-region
Setting an overlay for each different symbol and for the whole buffer
was terribly inefficient and didn't work for hidden regions which
didn't have a face yet.

Use fontify-region so we can take advantage of just in time
fontification and only darken region as soon as they become
visible/available for fontification.

Not sure this is the best approach, still working on it.