Handle face list properties
Still looking into this, sometimes get-text-property can return a list
of faces (e.g. when they are added with add-face-text-property). Take
the first one of the list when this happens.

So far with C modes I only saw it with #ifndef that has both
font-lock-preprocessor-face and font-lock-negation-char-face in the n
letter.