Only run our fontification hook if our mode is enabled
I'm still very doubtful that hooking into the default
fontify-region-function is the best approach as we're messing with
global fontification while being a buffer local minor mode.

Problem is c-mode extends the fontification region and chains into the
default function so if we want to get the extended region bounds we
need to hook into that. Maybe I should limit this ugly workaround to c
mode as ts modes don't seem to do this.